Subsonic: Optimize getting large folder contents
Before the optimization, fetching ~2000 tracks from a folder took 10-12
seconds on the test setup. After the optimization, it takes ~4 seconds.

Before the optimization, the Subsonic Music Streamer on Android
couldn't fetch the afore mentioned folder, apparently because it timed
out waiting the response. After the optimization, the folder shows up
fine on the said app.

+++ appframework/businesslayer/businesslayer.php
@@ -47,7 +47,9 @@ abstract class BusinessLayer {
 	 * @param array $ids the ids of the entities which should be deleted
 	 */
 	public function deleteById($ids) {
-		$this->mapper->deleteById($ids);
+		if (count($ids) > 0) {
+			$this->mapper->deleteById($ids);
+		}
 	}
 
 	/**
@@ -76,7 +78,11 @@ abstract class BusinessLayer {
 	 * @return Entity[]
 	 */
 	public function findById($ids, $userId=null) {
-		return $this->mapper->findById($ids, $userId);
+		if (count($ids) > 0) {
+			return $this->mapper->findById($ids, $userId);
+		} else {
+			return [];
+		}
 	}
 
 	/**
